refactor: 와 을 이용하여 구조를 간결하게 수정
