+++ src/components/FolderItem/FolderItem.js
@@ -2,7 +2,7 @@ import React from "react";
 
 function FolderItem({ id, name, onFolderClick }) {
   return (
-    <div className="folder-item" onClick={() => onFolderClick(id, name)}>
+    <div className="folder-item" onClick={() => onFolderClick(id)}>
       {name}
     </div>
   );
